+++ src/Modules/Admin/Widgets/Navs/SystemNavItem.php
@@ -19,6 +19,7 @@ class SystemNavItem extends NavItem
         $this->icon ??= 'cogs';
         $this->order ??= 999;
         $this->url ??= ['/admin/system/index'];
+        $this->roles ??= [self::ROLE_AUTHENTICATED];
 
         parent::__construct($config);
     }